dfsf

+++ commons/pkg/domain/eoditem.go
@@ -32,7 +32,7 @@ type EndOfDaySourceItem struct {
 	High      float32    `json:"high"`
 	Low       float32    `json:"low"`
 	Close     float32    `json:"close"`
-	Volume    int64      `json:"volume"`
+	Volume    float32    `json:"volume"`
 	AdjHigh   float32    `json:"adj_high"`
 	AdjLow    float32    `json:"adj_low"`
 	AdjClose  float32    `json:"adj_close"`
